상품이미지/이미지외파일업로드/ui처리완료

+++ ex02/src/main/java/com/spring/test/domain/BoardAttachVO.java
@@ -10,6 +10,6 @@ public class BoardAttachVO {
 	private String fileName;
 	private boolean fileType; //이미지 파일 여부를 판단 
 	
-	private int bno;
+	private long bno;
 	
 }
